+++ apps/frontend/src/components/admin/all-users/AllUsers.tsx
@@ -3,7 +3,7 @@ import { ArrowUpDown } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import AllUsersTable from "@/components/admin/all-users/AllUsersTable";
 import { useQueryState } from "@/hooks/useQueryState";
-import { UserAccountListType } from "@/types/admin";
+import { UserAccountListType } from "@/types/admin/user";
 
 const AllUsers = memo(({ searchText }: { searchText: string }) => {
   const [allUserQuery, setAllUserQuery] = useQueryState<UserAccountListType>(
