+++ apps/frontend/src/components/admin/books/CreateBook.tsx
@@ -4,9 +4,9 @@ import BookForm from "@/components/admin/books/BookForm";
 import { toast } from "sonner";
 import BackButton from "@/components/common/BackButton";
 import { useApi } from "@/hooks/useApi";
-import { BookFormType } from "@/types/admin";
 import { defaultBookValues } from "@/utils/admin/book";
 import { getQueryClient } from "@/utils/api";
+import { BookFormType } from "@/types/admin/book";
 
 const CreateBook = () => {
   const bookForm = useForm<BookFormType>({
