+++ apps/frontend/src/components/user/register/RegisterDetailsStepForm.tsx
@@ -1,7 +1,7 @@
 import { UseFormReturn } from "react-hook-form";
 import { CommonInput } from "@/components/form/CommonInput";
 import ImageUpload from "@/components/form/ImageUpload";
-import { UserRegisterType } from "@/types/userRegister";
+import { UserRegisterType } from "@/types/user/register";
 
 const RegisterDetailsStepForm = ({
   userRegisterForm,
