+++ apps/frontend/src/types/admin/user.ts
@@ -1,0 +1,10 @@
+import { UserAccountStatus } from "contract/enum";
+
+export type UserAccountListType = {
+  sortInAsc: boolean;
+  currentPage: number;
+};
+
+export type IdPreviewDataType = { idCardUrl: string; fullName: string };
+
+export type ChangeStatusModalType = { id: string; status: UserAccountStatus };
